psycho: update CMake script
It seems that CMake 3.16.3 doesn't know what the c_std_17 feature is.

C11 and C17 are virtually identical outside of defect fixes, and most
but not all of those defect fixes are relevant to us.

So, we use C17 if the CMake version is >=3.21.7. This appears to be the
first version that knows what c_std_17 is. Otherwise, we use c_std_11.

In both cases, compiler extensions are enabled.

Signed-off-by: lunaspis <lunaspis@proton.me>