psycho: disable some default compiler flags
Some Linux distributions (in this case, Ubuntu) set default compiler
flags to assist with security concerns. These flags will probably cause
performance problems and are mostly redundant in the presence of
sanitizers.

There's more work to be done on this front, but this is just a temporary
stopgap. I'm sure Ubuntu forces nonsense by default.

Signed-off-by: lunaspis <lunaspis@proton.me>